--- controllers/os-coreos-alibaba/charts/os-coreos-alibaba/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-appVersion: "1.0"
-description: A Helm chart for the Gardener CoreOS Container Linux extension (Alibaba's CoreOS)
-name: os-coreos-alibaba
-version: 0.1.0
