--- controllers/os-coreos-alibaba/charts/os-coreos-alibaba/templates/serviceaccount.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: gardener-extension-os-coreos-alibaba
-  namespace: {{ .Release.Namespace }}
-  labels:
-    app.kubernetes.io/name: gardener-extension-os-coreos-alibaba
-    helm.sh/chart: gardener-extension-os-coreos-alibaba
-    app.kubernetes.io/instance: {{ .Release.Name }}
