--- controllers/os-coreos-alibaba/example/controller-registration.yaml
@@ -1,16 +1,0 @@
----
-apiVersion: core.gardener.cloud/v1alpha1
-kind: ControllerRegistration
-metadata:
-  name: os-coreos-alibaba
-spec:
-  resources:
-  - kind: OperatingSystemConfig
-    type: coreos-alibaba
-  deployment:
-    type: helm
-    providerConfig:
-      chart: H4sIAAAAAAACA+1a7XPiNhrfz/4rnks/3O7NIgMJyZWb+0AJbZlLSSak29lPO8IIo4st+SQZQtO9v/0eyQYcnEvKlmZnW/2GwbYsPW/S8yKB1I1IKiZ1gyZ8Qic0fHVwNBFnnY67Inav7r51fNJqd9qnp7a9ddw87ryCzuFFqSPXhiqAV0pK81S/595/oZC1+e/PqTJkRdPkUDyem/92p7Uz/ydnp8evoPkSBviTzz/N+DumNJeiC4tWQLNs89gkbdIMpkxHimfGNfXge5akENklAjOpwMwZfEfVlAmmoI8L6XKMF2Eotw0XXOR3wO4ME5YkvO4Va+yvOoCy95tA0JR1obYOg8VDOT63of6gqPv/giY504cMAM/6//Hpjv932u0T7/8vAZ7SmHXRHRXLpOZGqlUXWE7iSBEuw7j07Uam5L9ZZDYN2zcb927MVxlTSMrQuAsJNUwbfMryJLmSCY+Q8HA2kuZKMc2ECQK8ylxFTHfh/mMQRFJEuVL4arwSETZ2vNP/7qj7v2Fp5ibvYJXg/vVfp3l87Ou/l8BT8z9lWSJXKTrkb8sGz8x/+6Rz9nD+281Wq+3j/0ugWv9h8adDLAJvuZh24Xwz+0HKDJ1SQ22eKMq1R6J/vYIrOuuMRjji/h7INUsY1YyM1s3w8SP2SuiEJdoSBysDuc0nTAmGS9CmoL0YAsyxQCV6HroadY9xdcZc4NIQj8luxdYZi8q8ibmN6qLXu6J4Khv7Mhem0FHj8Aiza6FlSk00v6iofQDF91cBM3Xp66VUlWm2SB4IeBARP0VINF5pa3fP1IJHrBdF1rijfQWI1puTjV6N/ZZ0gaJsgqPKnLsmsi2jUPhu7TWWRth+9JDOVaVCqo3Ylk9raxRqpClFL900NOCpimzdp6ZQw9pDySTBYUgQGVa7N7DtrrGtyxqK2QeeMP3PiqA7hVtVTiYWVRkLS18MeueD6w+Di0H/Zng5+jDq/TAYX/X6g01PALcL+VbJtFtpBJhxlkyv2exha9l+Rc28u1nFZBN8Nn23BWdl+P19A4x8jwmu4r5lP/gFBAZD1AxabavX5w7XB8dT+V9NaHSIfeBz+b/T3j3/OT09bvn8/xJoNBpBtQZwc05zM5eK/0ztqQ+5/bsL05vCoJ+gzZi6lgn79Mrgi8z5Kk9s7GjgQP6dknnmFGhsT7g0WfMmUSLzabATczACY0BGs4pYr9CIKYbOGY/1E69CFMjktseCqUlJJWbGXROui5ulLSjcXba5yzOcGVaX9uioLpZmkWLm13PB3pZ2hc2W969iWGiX0qzQnS0wxu6wL3nsT279zpWZpdlriQ9ndmqXiC3LcOYe13y5a8yKlr/Jcb7BBpzpP5v/oOZl8l7P6ROGw171gLOXmXQ+sYdWzmULUuMH1ePvsZ353AF9TzyV/8tSmxbG+vRK4Lnzn5Oz5u7+/+yk6fP/S2Dn959HvcRv/+uh7HPP26FQ939iTchjga0H4vGc/6Pb757/ts463v9fAl8B7poxuQqNm2Aoph2WcyZgkvPE1ihY70S3NEanCL6CmznXoPMsk8rgDS6VBOJETooTLez9FqsvzB58wWydNK+0UzFFAoLFLsfD60yxGb9jU1hy7PeXNwQuRbICKdxIKxJgQYzFp2AkIOfjD2ODsiGJvkxTJPCuP4YpVzogMTeh+y7ED8jkZxW673XDPA7t1/pRL0S4JYT1x22ewcweawR/I3qZ4feE3uK3SfH+v9j1HVVc5hqG5wNkWP4YFhA+ZTQs+mFTQBY6klMWfkHRoe7/UxmRWB6Sx3P7f/zs/v+jhSHB+/8LIAzRDbIVesrcwOvoDbSbra9h3LuC8QCkQq91D3SG7sGxKrQnjxkVKwI9dH03TNsNF1YMbEqK+GA9CfCaYA0hNHp4LnCr5f4r0sNggpexnJklxUhzUXR5CwsCbdxHRywzQDUIaXCcxCFqyTVSE274xbA/GKFglkMQhvhZU3iEyYZ2WeBAmzThte1wVL46evMPS2Ilc4xTK8sUcmRmNkqUAiF3qzYaAOuBIl6ZLQNiabwvaciJPV4GigMyfJpVOwI1pdAOc2Oybhgul0tCncREqjgsjabDUteG/fNLMepHgRHKWvs/OVeo8WRlyxb7Q8MEZU3o0k1YrBi+s8FcwFJx44KvLg1uyUxxO6/4JDcPjLaWEVWvdkCz4RI46o1hOD6Cb3rj4fitJfLT8Ob7yx9v4Kfe9XVvdDMcjOHyGvqXo/OhPdPFp2+hN3oP/xqOzt8C43Ym0ZwY9FEDFJNbc+KKsbTGjD0QYf3HInvsz2c8QtVEnGMKgljiJl24pMRUyrU7c3GZBckkPOXGJRdd14sE2CWW3dhWg3YdExJuPnPMAOH6TfVIXLHY2qLYlep5/RgBLteHNmN3aNN35xCw04mU3NgdRbVZ+P842B2WlRNzsku55R+tmLATrKEqfpmDna3KRmsWqzEyV5idYMsFHnAJsir1LyhXeXh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eGzxP/4hwZ8AUAAA
-      values:
-        image:
-          tag: 0.2.0
