+++ controllers/os-coreos-alicloud/charts/os-coreos-alicloud/.helmignore
