+++ controllers/os-coreos-alicloud/charts/os-coreos-alicloud/Chart.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+appVersion: "1.0"
+description: A Helm chart for the Gardener CoreOS Container Linux extension (Alicloud's CoreOS)
+name: os-coreos-alicloud
+version: 0.1.0
