+++ controllers/os-coreos-alicloud/charts/os-coreos-alicloud/templates/rbac.yaml
@@ -2,10 +2,10 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: gardener-extension-os-coreos-alibaba
+  name: gardener-extension-os-coreos-alicloud
   labels:
-    app.kubernetes.io/name: gardener-extension-os-coreos-alibaba
-    helm.sh/chart: gardener-extension-os-coreos-alibaba
+    app.kubernetes.io/name: gardener-extension-os-coreos-alicloud
+    helm.sh/chart: gardener-extension-os-coreos-alicloud
     app.kubernetes.io/instance: {{ .Release.Name }}
 rules:
 - apiGroups:
@@ -42,7 +42,7 @@ rules:
   resources:
   - configmaps
   resourceNames:
-  - os-coreos-alibaba-leader-election
+  - os-coreos-alicloud-leader-election
   verbs:
   - get
   - watch
@@ -52,16 +52,16 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: gardener-extension-os-coreos-alibaba
+  name: gardener-extension-os-coreos-alicloud
   labels:
-    app.kubernetes.io/name: gardener-extension-os-coreos-alibaba
-    helm.sh/chart: gardener-extension-os-coreos-alibaba
+    app.kubernetes.io/name: gardener-extension-os-coreos-alicloud
+    helm.sh/chart: gardener-extension-os-coreos-alicloud
     app.kubernetes.io/instance: {{ .Release.Name }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: gardener-extension-os-coreos-alibaba
+  name: gardener-extension-os-coreos-alicloud
 subjects:
 - kind: ServiceAccount
-  name: gardener-extension-os-coreos-alibaba
+  name: gardener-extension-os-coreos-alicloud
   namespace: {{ .Release.Namespace }}
