+++ controllers/os-coreos-alicloud/charts/os-coreos-alicloud/templates/serviceaccount.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: gardener-extension-os-coreos-alicloud
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app.kubernetes.io/name: gardener-extension-os-coreos-alicloud
+    helm.sh/chart: gardener-extension-os-coreos-alicloud
+    app.kubernetes.io/instance: {{ .Release.Name }}
