+++ controllers/os-coreos-alicloud/example/controller-registration.yaml
@@ -1,0 +1,16 @@
+---
+apiVersion: core.gardener.cloud/v1alpha1
+kind: ControllerRegistration
+metadata:
+  name: os-coreos-alicloud
+spec:
+  resources:
+  - kind: OperatingSystemConfig
+    type: coreos-alicloud
+  deployment:
+    type: helm
+    providerConfig:
+      chart: H4sIAAAAAAAAA+1a3XPbNhL3M/+KPefhkk5E6tutbu5BlZVWU5/sMd108pSBSIjCmQRYAJSs+nJ/exckJVMfteNGVSYpfqMRSXCxH1hgdwFJqFogJBWqRmIWxCILvZNDo44463SKa7e41pvt4lrgpNFqN5qdxlmnWz+pN1qts9YJdA6uyR5kShMJcBLWu51v29/9IR2STafHUOi4ELv+H8yI1O6SJPGBZDzl/2a3ueX/TquJ/q8fSP6j+Jv7n6TsLZWKCd6DecMhabp+rLstt14L6dwJqQokS3Xe3IcfaZxAYGYJTIUEPaPwA5Eh5VTCACfTpY8XrgkzDReMZ3dA7zTlhi287MdsQibkn8qBkvqVw0lCe7A7F535rjKfe8S+LuxZ/+4M/csijs2HkfHU+u80u9vrv9Oy6/8oeAFXRGsquQItoPA6LGaUwyRjcch4BCkJbklEleu8gJsZU6CyNBVS4w3OlBiiWEwgITqYIfVrkDQmms0p9tOzSjvhITLgNCI6jwOppFN2R0NYMKT7xysXLnm8BMHznkYlSDF+xBhFXMc999/7GnVDFgORJMjg7cCHkEnluBHTXv5dqO+4k9+kl3+vGmaRZ75Wj2rOvQdGE7QvS2HKYqqcb1y1SPF7Qm7xWyd4/38kfUskE5mC0fkQBaZS/JcG2nFZSIlX0GGT485VIELqfUExas/6D0XgRuKAMp5Y/816vb21/tvtbteu/2PA83AZpEtcKTMNL4NX0Kw3vgO/fwX+EDC5E54/oO0sZkRTCESSEr50oY9LP++mcMkrKuc0dIv4YFYS4BUnFCZ9XOEZD2lRJ/QxmODFF1O9IBhpLgqS1zB3oYlVQkBTDUQBFxr7CewiF0whN553vxgNhmNUzEhwPA8/Kw57hKx5lxUNNN06vDQEp+Wr01f/MiyWIsM4tTRCIUNhem1EqRBKN2bjAPCAFvFKPwhwDY93JQ8xMXUPEOyQ4tO0SghEl0rnmGmd9jxvsVi4JNfYFTLyykFTXmlrDbUue/3MMUKZ0f41YxItniwB6zXsQCaoa0wWucMiSfGdCeYcFhKDogm+qhxwwyZkSks2yfTGoK10RNOrBDhsOAVO+z6M/FP4vu+P/NeGyS+jmx8vf76BX/rX1/3xzWjow+U1DC7H56Ob0eUYn95Af/wOfhqNz18DZcaTOJwY9NECVJOZ4cQZY3j5lG6osCoqVUoDNmUBmsajDFMQRAILQp4nJSoTpoxbVZ5ZkE3MEqbz5KJ27XIdJIlELzJVqpnHruutPzPMAN7qDcZDrqWIYyprkkZmLHKmrprtqVDhMjW9UCV/qTRNsPCdsgi2qdxSHr0jaDj1/kiG2XUZTTEr50m3LLMpNy5WUDWgzML5aJWNZmCMzShdYn6CBymwIcVJq9y/oGxlcWjsyf84i1Ms4TAEHUjG889/Oq2zhj3/OQYe9X9I01gsE8o/7Tjoqfqv3T3b9H+z3u40bP13DFTPf7CYUN684dwyHvbgfO18J6GahESTngNQHNVE5XlPbX2wU9tzelNQK8w22OX+HtxrGlOCyXi8aoYPH5AqJhMaK8MdjBLubTbBNE9xCrpMeM+TiHUVjRNM1l6e3J7TcVc04+h2vk97o7ipT4zSkuZFmCqo3pI4w95l40BkXBdWKuwe4B62sDPfGF9UDD+E6c+3AWC13Eu1Kq42iDc0PIyOf0ZLgNVo5/e43cC6rh8EZnjHz9YgWB1Pri2rPXNiF2AJVlE9OK34PW8y3heKobOXqH5v57UmEbafbvK5yuL4SqCI5cZEKnqk65er8SjsSBKsfh/cUwNvjwWzJRapFZpdi6pFL3JEiVX6GrbdGZIgw8KSa6xZzYM5MPl3RdMHAn/JA1VVlPJ5VclirC+G/fPh9fvhxXBg9gzvx/3/DP2r/mC4pgSYG9ZvpEh6lUbA3RmNw2s63Wwt26+InvXWM9ldx6A1LW5CRCYDqqrd7+9ruGt6h3musoZLOvgf4K4Dh1VDo2ns+txh+2B4NP/LCQkO8EPQU/m/09r+/eescWbz/1FQq9Wcag2Qu5xkeiYk+63YLd5+m0fodWEwiHHMqLwWMf2EyuALzfkyi03YqGFH9oMUWZqbUHv4gUu5K+HuSspGuMHouzoxUPmJQZCfGKhHXnmokM4MxZzKScklojq/xritz28WpqDI79L1XZaic+iutqenu2opGkiqP14KUhveFTEPsj9KYGFdQtLCdjrH8LolvpTxfHard3mhWQ77btJD14Zmkpi6DF233/TF9mhWzPykxfM9NqCr/4ZrCI0vc/fKr4+MHVLtxp3njZTKJub3mnzdFrz8jQLyL9nWfO7I/nF4NP+XZTYpRulPVwJPnf+0z1pb+/9Gq27z/1Gw9f+PvavDbv/3hrHP7bqDYM/6nxebn8P9AezJ/3+1dv//0W3a9X8MFCcYxSFWeWLRA5q5USDNClivnvIvD+uGx44ZNIl6kGcQk1nTyrHGaDoW+sr8XIxhxXko4OD+g+NsHR/0oPN1LDELCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLC4u/BL8DCfdEvQBQAAA=
+      values:
+        image:
+          tag: 0.3.0-dev
