+++ controllers/os-coreos-alicloud/pkg/coreos-alicloud/internal/cloudinit/cloudinit.go
