+++ controllers/os-coreos-alicloud/pkg/coreos-alicloud/internal/generator.go
