+++ controllers/os-coreos-alicloud/pkg/coreos-alicloud/internal/template_generator.go
