+++ controllers/os-coreos-alicloud/pkg/coreos-alicloud/internal/templates/cloud-init.sh.template
