+++ controllers/os-coreos-alicloud/pkg/coreos-alicloud/internal/testfiles/cloud-init.sh
