+++ controllers/os-coreos/example/controller-registration.yaml
@@ -10,7 +10,7 @@ spec:
   deployment:
     type: helm
     providerConfig:
-      chart: H4sIAAAAAAACA+1aW3PbNhbOM3/FWfcl6USgJVvyrjr7oMpKq6kreyw3nTxlIAqisCYBFgAlq2762/cApChKdmOn8WWzwTcakQTBcwPOBSClbkRSManDF4+GfcRRu+2OiN2jO28eHDZb7VanY9ubzcPO4QtoP55IG+TaUAXwQklpPtbvrvtfKGQ1/v05VYasaJo8NI+7xr/VPNwZf7xovoD9pzDAVz7+NONvmdJcii4smgHNsupyn7TIfjBlOlI8M66pBz+yJIXIThWYSQVmzuAHqqZMMAV9nEinYzwIQ7ltOOEivwJ2ZZiwJANBU9aFasoFi21Wz22LrxEb/1/QJGf6MQLAnf5/0Nn1/077wPv/U4CnNGbdAECxTGpupFp1geUkjhThMoxL325kSv6HRaZq2Nyp3LsxX2VMISlD4y4k1DBt8CrLk+RMJjxCwsPZSJozxTQTJgjwKHMVMd2F6w9BEEkR5UrhrfFKRNjY9hHh0bHxf8PSzA3ag1eCn17/HRx2Wr7+ewrcNv5TliVylaIjPkw2uGP8Wwfto53x77QP9338fwrU6z8s/rAMaAaXXEy7cFzNgiBlhk6poTZPFDXcLdF/U9YVnXRGI+x5fQ3knCWMakZG62b48AF7JXTCEm2JguVNLvMJU4LhFLSp516MAOZYkBI9D11Neo/+NxlxgVNA3CarFVNnLCrzI+Ywqoteb4tiqWzsy1yYQieNj0eYRQutUmqi+UlNzc9Q9NNFx0xc+nQpTW0YLZItwT5LtL8jHBqrtK07Z2rBI9aLImvM0X0ZR+vFRqVH435TtEBR/sBebUxdE9mUQyhs98ZtLHGwfW+bzlmt0rnxxKYMWmtfiJ+mFL2tamjAxyqrdZ9KkYbVX8kkwe5ICBnVuzWw7aqxqasaitkLnjD975qAO4VXXT4mFnXZCsueDHrHg/P3g5NB/2J4Ono/6v08GJ/1+oOqJ4BbTbxRMu3WGgFmnCXTczbbbi3bz6iZd6tZSqogUvXdFIy1x6+vG2DkO0xUNbcs+8EfIDCYoWbQbFm9njvc/s/htvyvJjR6yHXgXfnfbvZs5/92q93x+f8p0Gg0gnoN4Mae5mYuFf+d2l0fcvlPF86rwqCfoM2YOpcJ+/TK4IvI+SpPbIxp4IP8ByXzzAnc2OxkabLmSaJE5tNgJzZhhMaAjeYTsV6hsVIMsTMe64/cClEgk9seC6YmJZWYGXdMuC5OlragcGdZdZZnOALsprR7ezfF0ixSzNyfC/a2tGtsNrzvxbDQLqVZoTtbYCzeYV/y+HRy63uurCzNXiVGHNGpnRK2HLN7j7dqvNw1Yk27z3KM77EBR/j/1T9QwzKJr8fsIwbCXjcDx73MofOJ3XRyrliQGG9ViQ+5HHnuQPxMuC3/l6U4LYz8+ZXAXfs/h4e7+z9HB/b9n8//j4+d9z+3etnXvPx/7vF5bGz8n1hT8ljg1QPzuMv/0e1vvP9pef9/EnwDuOrGpCw0LqKhGH5YzpmASc4TW8NgPRRd0hidJfgGLuZcg86zTCqDJzhlEogTOSl2urD3a6zKMIvwBbN11LzWTsUUCQgWu9oAXmaKzfgVm8KSY79/vCJwKpIVSOGetCIBFspYlApGAnI8fj82KBuS6Ms0RQJv+2OYcqUDEnMTuv9C/IBMfleh+183zOPQ/q0v9UKEG0JYt1zmGczstkjwLdHLDP8n9BL/TYrnf2LXt1RxmWsYHg+QYfkyLCB8ymhY9MOmgCx0JKcs/IKixsb/pzIisXwMHnf4f/Oo09rx/1b7yH//8SQIQ3SDbIWeMjfwMnoFrf3mv2DcO4PxAKRCr3UXdIbuwbE6tDuWGRUrAj10ffeYtgsxrBjYlBTxwXoS4DHBGkJo9PBc4FLMfSvSw2CCh7GcmSXFSHNSdHkNCwItXF9HLDNANQhp8DmJj6gl10hNuMdPhv3BCAWzHIIwxN+awi1MKtplgQMtsg8vbYe98tbeq+8siZXMMU6tLFPIkZmplCgFQu5WbTQA1glFvDIbBsTSeFfSkBO7HQ0UH8jwalbvCNSUQjvMjcm6YbhcLgl1EhOp4rA0mg5LXRv2y5jiqV8ERihr7d9yrlDjycqWM/YFxARlTejSDVisGN6zwVzAUnHjgq8uDW7JTHGZr/gkN1tGW8uIqtc7oNlwCuz1xjAc78H3vfFw/NoS+XV48ePpLxfwa+/8vDe6GA7GcHoO/dPR8dDuCePVG+iN3sFPw9Hxa2DcjiSaE4M+aoBicmtOnDGW1pixLRHWHxbZ1wN8xiNUTcQ5piCIJS7ihUtKTKVcu70Yl1mQTMJTblxy0Tf1IgF2iWU3ttWhnceEhNVvjhkgXN+pb6krFltbFKtZPd9sL8DpehNn7DZx+m5fAsqbpKTOriiqycK/omhXVlYuzMEuxZYfVjFhB1RDXdwy5zrblI3WDFZDZKowG8GGC2xxCbI69S8oN3l4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHh4eHj8Nf4LONi7BQBQAAA=
+      chart: H4sIAAAAAAAAA+1aX3PbNhL3Mz/FnvuS3ESkJFtSq5s+qLLaauqTPaabTp4yEAlRqEmABUDJqi/32W9BUhQlq7HT2PK1wW88Igku9g8W2F2AFqoRCEmF8o6eDU1Er9Mprt3i2myfFtcCR62T01a70+p1us2jZqvVbbaOoPN8Km2QKU0kwFHY7Ha+Pv3mD+mQbDY7hEKHhaj8P5wTqd0VSeKnlvGQ/9t4v+3/k067dwTNp1ZkH75w/5OUvaVSMcH7sGg5JE2rx6Z74jYbIV04IVWBZKnOmwfwI40TCMx0gZmQoOcUfiAypJxKGOJkuvDxwjVhpuGc8ewW6K2m3LB1OEloH6pp5yzui3vpMfmSsFn/7hzdyiKOT08s46H132l3d9d/77Rl1/8h8BVcEq2p5Aq0gML9sJxTDtOMxSHjEaQkuCERVa7zFVzPmQKVpamQGm9wysQQxWIKCdHBHKnfgKQx0WxBsZ+e19oJD5EBpxExcQRepZLO2C0NYcmQ7h+vXbjg8QoEz3salSDF+BFjFHEd98x/72vUDVkMRZIgg7dDH0ImleNGTHv5b6G+405/l17+u26YR575WT+qBfc2jKZoX5bCjMVUOf901TLF3ym5wV+d4P1/kfQtkUxkCsZnIxSYSvErDbTjspASr6DDJsddqECE1PsLRbDN+g9F4EbiOWQ8sP5bvW57Z/23e82uXf+HgOfhMkhXuFLmGl4Fr6HdbH0D/uAS/BFgcic8f0DbWcyIphCIJCV85cIAl37eTeGSV1QuaOgW8cGsJMBrzAJM+rjCMx7Sok4YYDDBiy9mekkw0pwXJG9g4UIbq4SAphqIAi409hPYRS6ZQm48734+Ho4mqJiR4Hge/q057BFS8S4rGmi7TXhlCI7LV8ev/2VYrESGcWplhEKGwnRlRKkQSjdm4wDwgBbxSm8EuIbHu5KHmJq6Bwh2SPFpVicEokulc8y1Tvuet1wuXZJr7AoZeeWgKa+0tYFal71+5hihzGj/ljGJFk9XgPUadiBT1DUmy9xhkaT4zgRzDkuJQdEEX1UOuGETMqUlm2Z6a9DWOqLpdQIcNpwCxwMfxv4xfDfwx/4bw+SX8fWPFz9fwy+Dq6vB5Ho88uHiCoYXk7Px9fhigk/fw2DyDn4aT87eAGXGkzicGPTRAlSTmeHEGWN4+ZRuqbAuKlVKAzZjAZrGowxTEEQCy0WeJyUqE6aMW1WeWZBNzBKm8+Si7tvlOkgSiX5kqlQzj13Xq/7mmAG89RuMh1xLEcdUNiSNzFjkTF0135StcJEaYtTEXylNE6x3ZyyC8qVbcqe3BM2k3h9xNJstoxfm4DzFlkU15cahCurqljk3H5uy0QyDsRCFSsxGsJECW1KctM59Jzdt4j9akWLmpk9/EvDp+/+TTq9n9/+HwD7/hzSNxSqh/ImOAx7wfxu9veP/bhfJbf4/AOr7f0wmylu0nBvGwz6cVbPASagmIdGk7wAU+/eo3O83qo19Y7OlL4gUhh2kvLsD94rGlGAMnqyb4cMHpIrJlMbKMAUj273JphjdKU5BlwnvUYIwi+KuFUOzlwe3R9DfF8Q4+pbv09WoaZKQUVHSPNOqguotiTPsXTYORcZ1YZPC7gFuVAqr8t3Pec3MzzD001UHWK/pUpuaGw3iLcU+S7U/oxzAemzze6wgMVUPgsAM5uSxgoP1QVNlR+NxU7QASzAx9uG45tO8yXhWKIaOXKGy/XuvNYmw/Xibz2UWx5cCJ8Rqa5IUPdLq5dr6Qv0kwfJl44MGeHsUn6+w3KjRVIbUixVkhILqZA1suzUkQYYlAtdYfZgHs9H9tqbghsBf8UDV9aN8UdetGNnz0eBsdPV+dD4amlrv/WTw75F/ORiOKkqAhWH9vRRJv9YIWFXTOLyis+3Wsv2S6Hm/mqVuFUQqWiweRSYDqurd7+4aWO2+w0RVW5YlHfwHsFrE0dTQahu7Xjrc/t9hX/6XUxI85YeAh/L/aau1k/87J70Tm/8PgUaj4dRrgNz3JNNzIdnvxf7h5us8nFeFwTDGMaPySuAe/JMrg79EzpdZbGJMAzuyH6TI0lzhxuYrhnLXMt0gFlno7MQmjNDr/aHK94dBvj9UH3nloUI6MxS4xZ2WXCKq82uMu7n8ZmkKivwure6yFD1A72t7fHxfLUUDSfXjpSC14V0Ts5H9KIGFdQlJC9vpAmPxjvhSxqezW7/Ly8py2KvEiB4NzZQw5Zj57rTX4uXuINas+6yF8R02oIf/rusDLSyT+NpnHxkgpLofOB41HCqbmnP2fCkWLPytKvEptyMvHYhfCPvyf1mKk2KQP78SeOj85/R09/yn12na738Hwc73/72r7Eve/r+0f54bm/W/KPZOz/APQA/+/8/J/e//vaZd/4dAcQBSnG+VBx59oJkbBdIsjmo9lZ+8q4aPnVJoEvUhTyUmQ6e1U5HxbCL0pflciGHF2dR2cPfBcXaOIfrQ+duvPgsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwsLCwuLP4f/AY9VPcYAUAAA
       values:
         image:
-          tag: 0.2.0
+          tag: 0.3.0-dev
