+++ extensions.yaml
@@ -4,6 +4,6 @@ extensions:
 - name: os-coreos
   gitHubRepo: https://github.com/gardener/gardener-extensions
   path: controllers/os-coreos
-- name: os-coreos-alibaba
+- name: os-coreos-alicloud
   gitHubRepo: https://github.com/gardener/gardener-extensions
-  path: controllers/os-coreos-alibaba
+  path: controllers/os-coreos-alicloud
